backend/requirements.txt
opencv-python==4.8.1.78
ultralytics==8.0.196
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.1
imutils==0.5.4